Added isolated storage handling
