+++ Commands/Provisioning/Tenant/SaveTenantTemplate.cs
@@ -4,6 +4,7 @@ using OfficeDevPnP.Core.Framework.Provisioning.Model;
 using OfficeDevPnP.Core.Framework.Provisioning.Providers;
 using OfficeDevPnP.Core.Framework.Provisioning.Providers.Xml;
 using SharePointPnP.PowerShell.CmdletHelpAttributes;
+using SharePointPnP.PowerShell.Commands.Utilities;
 using System;
 using System.IO;
 using System.Linq;
@@ -85,25 +86,8 @@ namespace SharePointPnP.PowerShell.Commands.Provisioning.Tenant
 
             if (extension == ".pnp")
             {
-                var useNewEvidence = false;
-                try
-                {
-                    var usfdAttempt1 = System.IO.IsolatedStorage.IsolatedStorageFile.GetUserStoreForDomain(); // this will fail when the current AppDomain Evidence is instantiated via COM or in PowerShell
-                }
-                catch (Exception e)
-                {
-                    useNewEvidence = true;
-                }
 
-                if (useNewEvidence)
-                {
-                    var replacementEvidence = new System.Security.Policy.Evidence();
-                    replacementEvidence.AddHostEvidence(new System.Security.Policy.Zone(System.Security.SecurityZone.MyComputer));
-
-                    var currentAppDomain = System.Threading.Thread.GetDomain();
-                    var securityIdentityField = currentAppDomain.GetType().GetField("_SecurityIdentity", System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.NonPublic);
-                    securityIdentityField.SetValue(currentAppDomain, replacementEvidence);
-                }
+                IsolatedStorage.InitializeIsolatedStorage();
 
                 var templateFileName = outFileName.Substring(0, outFileName.LastIndexOf(".", StringComparison.Ordinal)) + ".xml";
