+++ Commands/SharePointPnP.PowerShell.Commands.csproj
@@ -788,6 +788,7 @@
     <Compile Include="Utilities\CertificateHelper.cs" />
     <Compile Include="Utilities\Clipboard.cs" />
     <Compile Include="Utilities\FileUtilities.cs" />
+    <Compile Include="Utilities\IsolatedStorage.cs" />
     <Compile Include="Utilities\OperatingSystem.cs" />
     <Compile Include="ClientSidePages\WebPartMappingLoader.cs" />
     <Compile Include="Utilities\REST\MetadataType.cs" />
